requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langgraph>=0.2.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pdfplumber>=0.11.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-docx>=1.0.0
tzdata>=2024.1
